+++ common/src/main/java/net/mehvahdjukaar/hauntedharvest/HauntedHarvest.java
@@ -8,6 +8,7 @@ import net.mehvahdjukaar.hauntedharvest.configs.CommonConfigs;
 import net.mehvahdjukaar.hauntedharvest.integration.CompatHandler;
 import net.mehvahdjukaar.hauntedharvest.network.NetworkHandler;
 import net.mehvahdjukaar.hauntedharvest.reg.*;
+import net.mehvahdjukaar.moonlight.api.entity.VillagerAIHooks;
 import net.mehvahdjukaar.moonlight.api.misc.EventCalled;
 import net.mehvahdjukaar.moonlight.api.platform.ClientHelper;
 import net.mehvahdjukaar.moonlight.api.platform.PlatHelper;
@@ -28,6 +29,7 @@ import net.minecraft.stats.Stats;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.sensing.NearestBedSensor;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ArmorItem;
@@ -65,6 +67,7 @@ public class HauntedHarvest {
         ModRegistry.init();
         ModTabs.init();
         CompatHandler.init();
+        HalloweenVillagerAI.init();
         if (PlatHelper.getPhysicalSide().isClient()) {
             ClientRegistry.init();
             ClientHelper.addClientSetup(ClientRegistry::setup);