+++ common/src/main/java/net/mehvahdjukaar/hauntedharvest/client/model/HalloweenMaskModel.java
@@ -1,4 +1,4 @@
-package net.mehvahdjukaar.hauntedharvest.client;
+package net.mehvahdjukaar.hauntedharvest.client.model;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;