+++ common/src/main/java/net/mehvahdjukaar/hauntedharvest/items/PaperBagItem.java
@@ -3,6 +3,8 @@ package net.mehvahdjukaar.hauntedharvest.items;
 import dev.architectury.injectables.annotations.PlatformOnly;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.monster.EnderMan;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Block;
@@ -17,5 +19,10 @@ public class PaperBagItem extends BlockItem {
     @PlatformOnly(PlatformOnly.FORGE)
     public @Nullable EquipmentSlot getEquipmentSlot(ItemStack stack) {
         return EquipmentSlot.HEAD;
+    }
+
+    @PlatformOnly(PlatformOnly.FORGE)
+    public boolean isEnderMask(ItemStack stack, Player player, EnderMan enderMan) {
+        return true;
     }
 }