+++ fabric/src/main/java/net/mehvahdjukaar/hauntedharvest/integration/fabric/FDCompatImpl.java
@@ -1,5 +1,6 @@
 package net.mehvahdjukaar.hauntedharvest.integration.fabric;
 
+import com.google.common.base.Preconditions;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.food.FoodProperties;
 
@@ -11,7 +12,7 @@ public class FDCompatImpl {
         return new FoodProperties.Builder()
                 .nutrition(12)
                 .saturationMod(0.8F)
-                .effect(new MobEffectInstance(NOURISHMENT.get(), 3600, 0), 1.0F)
+                .effect(new MobEffectInstance(Preconditions.checkNotNull(NOURISHMENT.get()), 3600, 0), 1.0F)
                 .build();
     }
 }