+++ forge/src/main/java/net/mehvahdjukaar/hauntedharvest/forge/HHPlatformStuffImpl.java
@@ -26,7 +26,4 @@ public class HHPlatformStuffImpl {
         }
     }
 
-    public static ShaderInstance getBlur() {
-        return HauntedHarvestForgeClient.getBlur();
-    }
 }