+++ forge/src/main/java/net/mehvahdjukaar/hauntedharvest/forge/HauntedHarvestForge.java
@@ -1,20 +1,11 @@
 package net.mehvahdjukaar.hauntedharvest.forge;
 
-import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.*;
 import net.mehvahdjukaar.hauntedharvest.HauntedHarvest;
 import net.mehvahdjukaar.hauntedharvest.reg.ClientRegistry;
 import net.mehvahdjukaar.hauntedharvest.reg.ModRegistry;
-import net.mehvahdjukaar.moonlight.api.client.texture_renderer.FrameBufferBackedDynamicTexture;
-import net.mehvahdjukaar.moonlight.api.client.texture_renderer.RenderedTexturesManager;
 import net.mehvahdjukaar.moonlight.api.platform.PlatHelper;
 import net.mehvahdjukaar.moonlight.api.util.Utils;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.Screenshot;
-import net.minecraft.client.gui.screens.inventory.MerchantScreen;
-import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.level.block.Blocks;
@@ -24,17 +15,10 @@ import net.minecraftforge.event.TagsUpdatedEvent;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
-import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.MissingMappingsEvent;
-
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.function.Consumer;
 
 /**
  * Author: MehVahdJukaar
@@ -48,10 +32,14 @@ public class HauntedHarvestForge {
 
         MinecraftForge.EVENT_BUS.register(this);
 
-        PlatHelper.addCommonSetup(()->{
+        PlatHelper.addCommonSetup(() -> {
             ((FlowerPotBlock) Blocks.FLOWER_POT)
                     .addPlant(Utils.getID(ModRegistry.CORN_BASE.get()), ModRegistry.CORN_POT);
         });
+
+        if (PlatHelper.getPhysicalSide().isClient()) {
+            HauntedHarvestForgeClient.init();
+        }
     }
 
     @SubscribeEvent(priority = EventPriority.HIGH)
@@ -67,9 +55,9 @@ public class HauntedHarvestForge {
 
     @SubscribeEvent
     public void onServerTick(TickEvent.ServerTickEvent event) {
-        if( event.phase == TickEvent.Phase.END){
+        if (event.phase == TickEvent.Phase.END) {
             ServerLevel overworld = event.getServer().overworld();
-            if(overworld.getGameTime() % 10000 == 0){
+            if (overworld.getGameTime() % 10000 == 0) {
                 HauntedHarvest.getSeasonManager().refresh();
             }
         }